[FIX] fixes for Python 2.6

+++ pyOpenMS/tests/unittests/test000.py
@@ -4021,5 +4021,5 @@ def testIndexedMzMLDecoder():
 
 def test_streampos():
     p = pyopenms.streampos()
-    assert isinstance(long(p), long)
+    assert isinstance(int(p), int)
